chore: upload prior to refactor/tear apart; squash
